Day 6 - Part Two (Still unhappy but fast enough...)

+++ src/test/java/it/lucadom/aoc2023/day6/RaceTest.java
@@ -2,14 +2,16 @@ package it.lucadom.aoc2023.day6;
 
 import org.junit.jupiter.api.Test;
 
+import java.math.BigInteger;
+
 import static org.junit.jupiter.api.Assertions.*;
 
 class RaceTest {
 
     @Test
     void wins() {
-        assertEquals(4, new Race(7, 9).wins());
-        assertEquals(8, new Race(15, 40).wins());
-        assertEquals(9, new Race(30, 200).wins());
+        assertEquals(BigInteger.valueOf(4), new Race(BigInteger.valueOf(7), BigInteger.valueOf(9)).wins());
+        assertEquals(BigInteger.valueOf(8), new Race(BigInteger.valueOf(15), BigInteger.valueOf(40)).wins());
+        assertEquals(BigInteger.valueOf(9), new Race(BigInteger.valueOf(30), BigInteger.valueOf(200)).wins());
     }
 }
