+++ src/test/java/it/lucadom/aoc2023/day6/RacesTest.java
@@ -2,6 +2,7 @@ package it.lucadom.aoc2023.day6;
 
 import org.junit.jupiter.api.Test;
 
+import java.math.BigInteger;
 import java.util.List;
 
 import static org.junit.jupiter.api.Assertions.*;
@@ -12,9 +13,9 @@ class RacesTest {
     void fromInput() {
         assertEquals(new Races(
                 List.of(
-                        new Race(7, 9),
-                        new Race(15, 40),
-                        new Race(30, 200)
+                        new Race(BigInteger.valueOf(7), BigInteger.valueOf(9)),
+                        new Race(BigInteger.valueOf(15), BigInteger.valueOf(40)),
+                        new Race(BigInteger.valueOf(30), BigInteger.valueOf(200))
                 )
         ), Races.fromInput(List.of(
                 "Time:      7  15   30",
